Shadow for PR #8177, allowing to see the diff between boogie-backend-exp and boogie-backend. This is achieved by simply copying boogie-backend-exp over boogie-backend. This PR is not compiling or testing.
