--- language/move-prover/boogie-backend-exp/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "boogie-backend-exp"
-version = "0.1.0"
-authors = ["Diem Association <opensource@diem.com>"]
-description = "Move prover Boogie backend"
-license = "Apache-2.0"
-publish = false
-edition = "2018"
-
-[dependencies]
-async-trait = "0.1.42"
-bytecode = { path = "../bytecode" }
-move-model = { path = "../../move-model" }
-boogie-backend = { path = "../boogie-backend" }
-vm = { path = "../../vm" }
-diem-workspace-hack = { path = "../../../common/workspace-hack" }
-num = "0.4.0"
-itertools = "0.10.0"
-move-core-types = { path = "../../move-core/types" }
-log = "0.4.14"
-serde = { version = "1.0.124", features = ["derive"] }
-serde_json = "1.0.64"
-anyhow = "1.0.37"
-once_cell = "1.7.2"
-pretty = "0.10.0"
-regex = "1.4.3"
-rand = "0.8.3"
-futures = "0.3.12"
-handlebars = "3.5.3"
-tokio = { version = "1.3.0", features = ["full"] }
-codespan = "0.8.0"
-codespan-reporting = "0.8.0"
