--- language/move-prover/boogie-backend-exp/src/prelude/multiset-array-theory.bpl
@@ -1,44 +1,0 @@
-// Copyright (c) The Diem Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-// Boogie model for multisets, based on Boogie arrays. This theory assumes extensional equality for element types.
-
-type {:datatype} Multiset _;
-function {:constructor} Multiset<T>(v: [T]int, l: int): Multiset T;
-
-function {:builtin "MapConst"} MapConstMultiset<T>(l: int): [T]int;
-
-function {:inline} EmptyMultiset<T>(): Multiset T {
-    Multiset(MapConstMultiset(0), 0)
-}
-
-function {:inline} LenMultiset<T>(s: Multiset T): int {
-    l#Multiset(s)
-}
-
-function {:inline} ExtendMultiset<T>(s: Multiset T, v: T): Multiset T {
-    (var len := l#Multiset(s);
-    (var cnt := v#Multiset(s)[v];
-    Multiset(v#Multiset(s)[v := (cnt + 1)], len + 1)))
-}
-
-// This function returns (s1 - s2). This function assumes that s2 is a subset of s1.
-function {:inline} SubtractMultiset<T>(s1: Multiset T, s2: Multiset T): Multiset T {
-    (var len1 := l#Multiset(s1);
-    (var len2 := l#Multiset(s2);
-    Multiset((lambda v:T :: v#Multiset(s1)[v]-v#Multiset(s2)[v]), len1-len2)))
-}
-
-function {:inline} IsEmptyMultiset<T>(s: Multiset T): bool {
-    (l#Multiset(s) == 0) &&
-    (forall v: T :: v#Multiset(s)[v] == 0)
-}
-
-function {:inline} IsSubsetMultiset<T>(s1: Multiset T, s2: Multiset T): bool {
-    (l#Multiset(s1) <= l#Multiset(s2)) &&
-    (forall v: T :: v#Multiset(s1)[v] <= v#Multiset(s2)[v])
-}
-
-function {:inline} ContainsMultiset<T>(s: Multiset T, v: T): bool {
-    v#Multiset(s)[v] > 0
-}
