--- language/move-prover/boogie-backend-exp/src/prelude/vector-smt-array-theory.bpl
@@ -1,11 +1,0 @@
-// Copyright (c) The Diem Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-// Boogie model for vectors, based on smt arrays.
-
-// This version of vectors requires boogie to be called with `-useArrayTheory`.
-// It is not extensional.
-
-// Currently we just include the basic vector array theory.
-
-{{> vector-array-theory }}
